src/requirements.txt
gradio==6.13.0
ollama==0.6.2
langchain-community==0.4.1
pymupdf==1.27.2.3
pywebview
